+++ controllers/homepage-routes.js
@@ -1,28 +1,7 @@
 const router = require('express').Router();
-// const sequelize = require('../config/connection');
 const { Story, Author, Chapter } = require('../models');
 const userAuth = require('../utils/userAuth');
-const natural = require('natural');
 const analyzeText = require('../utils/natural');
-const Analyzer = require('natural').SentimentAnalyzer;
-const stemmer = require('natural').PorterStemmer;
-const analyzer = new Analyzer("English", stemmer, "afinn");
-const tokenizer = new natural.WordTokenizer();
-
-// Function to analyze story sentiment
-const checkStory = (stories) => {
-    return new Promise((res, rej) => {
-        res(stories.map(story => {
-            const storyText = story.story_text;
-            const splitText = tokenizer.tokenize(storyText);
-            const analyzedText = analyzer.getSentiment(splitText);
-            return {
-                ...story,
-                storyScore: analyzedText
-            };
-        }));
-    });
-};
 
 // Route to get stories for homepage
 router.get('/', (req, res) => {
@@ -46,9 +25,8 @@ router.get('/', (req, res) => {
     })
         .then(storyData => {
             const stories = storyData.map(story => story.get({ plain: true }));
-            checkStory(stories)
+            analyzeText(stories)
                 .then(analyzerData => {
-
                     res.render('homepage', {
                         stories: analyzerData,
                         loggedIn: req.session.loggedIn
